+++ src/main/java/io/gatling/demostore/website/controllers/AdminProductsController.java
@@ -1,4 +1,4 @@
-package io.gatling.demostore.controllers;
+package io.gatling.demostore.website.controllers;
 
 import javax.validation.Valid;
 import java.io.IOException;
