+++ src/main/java/io/gatling/demostore/website/controllers/RegistrationController.java
@@ -1,4 +1,4 @@
-package io.gatling.demostore.security;
+package io.gatling.demostore.website.controllers;
 
 import javax.validation.Valid;
